requirements.txt
pandas
numpy
scikit-learn
mlflow
pyyaml
python-dotenv
streamlit
requests
joblib
pytest